bindings: Change return type of get_payload().
The function now returns a vector of bytes instead of a C-style byte
array for the C++ bindings. For the Python bindings the function now
returns a Python bytes object for Python version 3 and a Python
string object for Python version >= 2.6.

As Python version >= 2.7 is already required for the Python bindings
the requirements remain the same.