Simplify decoder API
The ability to limit the number of packets to be decoded as well as the
ability to count the number of decoded packets can be easily
implemented inside the decoder callback function.
Therefore, both features are removed from the decoder API to simplify
the API and the underlying algorithm as well.