Append Git revision hash to package version
Append the short Git revision hash of the current commit to the package
version if there is no release tag for the package version on it and Git
is available.
This makes the package version more informative and allows the
distinction between development and release versions.

This code is inspired by, but not derived from, a patch from Daniel
Elstner for the libsigrok project.